compression/requirements.txt
transformers
rouge-score
safetensors
py7zr
zarr
cupy-cuda12x
scikit-build
pytest
numcodecs
fastapi
uvicorn
triton
gekko
tokenizers
parameterized
pydantic
accelerate
ninja
loguru
datasets
sentencepiece
nvidia-ml-py
huggingface_hub
Cython
sacrebleu
evaluate
sqlitedict
peft
more_itertools
pytablewriter
autoawq
auto-gptq